Survey Questions: Add Type
When we create a question we'd like to have either a "Single" answer
or a "Multi" answer.

To do this we have to update the Queston Model to have a "type" field
and also update the form to handle selecting a type.

This does most of the work and adds a Request and Policy for future
use (when we start editing things)

This also fixes the "create Survey" policy which had some bad
logic in it.